[Cpp API Compatibility] Reuse cpp extension caster test
